+++ creatio_api_py/api/__init__.py
@@ -1,0 +1,4 @@
+from creatio_api_py.api.base import CreatioODataAPI
+
+
+__all__: list[str] = ["CreatioODataAPI"]
